# apps/vibe-coding-platform/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Vibe Coding Platform API
  description: API for AI-powered code generation and project management
  version: 1.0.0
  contact:
    name: Vibe Support
    url: https://github.com/your-repo/vibe-coding-platform
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://vibe.example.com/api
    description: Production server
...
paths:
...
  /projects/{id}/duplicate:
    post:
      tags: [Projects]
      summary: Duplicate a project
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '201':
          description: Project duplicated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '404':
          description: Project not found
  /projects/{id}/export:
    get:
      tags: [Projects]
      summary: Export project as JSON
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Project export file
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectExport'
  /projects/import:
    post:
      tags: [Projects]
      summary: Import a project from JSON
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectExport'
      responses:
        '201':
          description: Project imported
        '400':
          description: Invalid import format
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Project:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        sandboxId:
          type: string
        sandboxUrl:
          type: string
        template:
          type: string
        tags:
          type: array
          items:
            type: string
        isPublic:
          type: boolean
        isFavorite:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    ProjectExport:
      type: object
      properties:
        version:
          type: string
        exportedAt:
          type: string
          format: date-time
        project:
          type: object
          properties:
            name:
              type: string
            description:
              type: string
            template:
              type: string
            tags:
              type: array
              items:
                type: string
        files:
          type: array
          items:
            $ref: '#/components/schemas/GeneratedFile'
        conversations:
          type: array
          items:
            $ref: '#/components/schemas/Conversation'
    GeneratedFile:
      type: object
      properties:
        id:
          type: string
        path:
          type: string
        content:
          type: string
        language:
          type: string
    Conversation:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        messages:
          type: array
          items:
            $ref: '#/components/schemas/Message'
        createdAt:
          type: string
          format: date-time
    Message:
      type: object
      properties:
        id:
          type: string
        role:
          type: string
          enum: [user, assistant, system]
        content:
          type: string
        metadata:
          type: object
        createdAt:
          type: string
          format: date-time
...